Rating and completing Goal/subgoal/routine/task

+++ application/android/app/src/main/java/com/group12/beabee/models/responses/SubgoalDetail.java
@@ -13,7 +13,7 @@ public class SubgoalDetail implements Serializable {
     public int parentSubgoalId;
     @SerializedName("main_groupgoal_id")
     public int mainGroupGoalId;
-    public int rating;
+    public double rating;
     public String title;
     public String description;
     public boolean isDone;
